feat: add Aspect and Artemis producer

+++ src/main/java/com/saeed/minimalcms/repository/ComponentRepository.java
@@ -3,5 +3,5 @@ package com.saeed.minimalcms.repository;
 import com.saeed.minimalcms.entity.Component;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComponentRepository extends JpaRepository<Component, Long> {
+public interface ComponentRepository extends JpaRepository<Component, Integer> {
 }
