+++ src/main/java/com/saeed/minimalcms/repository/ContentRepository.java
@@ -3,5 +3,5 @@ package com.saeed.minimalcms.repository;
 import com.saeed.minimalcms.entity.Content;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContentRepository extends JpaRepository<Content, Long> {
+public interface ContentRepository extends JpaRepository<Content, Integer> {
 }
