+++ src/main/java/com/saeed/minimalcms/repository/ContentTypeRepository.java
@@ -3,5 +3,5 @@ package com.saeed.minimalcms.repository;
 import com.saeed.minimalcms.entity.ContentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContentTypeRepository extends JpaRepository<ContentType, Long> {
+public interface ContentTypeRepository extends JpaRepository<ContentType, Integer> {
 }
