+++ src/main/java/com/saeed/minimalcms/repository/FieldRepository.java
@@ -3,5 +3,5 @@ package com.saeed.minimalcms.repository;
 import com.saeed.minimalcms.entity.Field;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FieldRepository extends JpaRepository<Field, Long> {
+public interface FieldRepository extends JpaRepository<Field, Integer> {
 }
