+++ src/main/java/com/saeed/minimalcms/repository/WebHookHeaderRepository.java
@@ -3,5 +3,5 @@ package com.saeed.minimalcms.repository;
 import com.saeed.minimalcms.entity.WebHookHeader;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WebHookHeaderRepository extends JpaRepository<WebHookHeader, Long> {
+public interface WebHookHeaderRepository extends JpaRepository<WebHookHeader, Integer> {
 }
