+++ src/main/java/com/saeed/minimalcms/repository/WebHookRepository.java
@@ -3,5 +3,5 @@ package com.saeed.minimalcms.repository;
 import com.saeed.minimalcms.entity.WebHook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WebHookRepository extends JpaRepository<WebHook, Long> {
+public interface WebHookRepository extends JpaRepository<WebHook, Integer> {
 }
